Refactor #43 : 코드 리팩토링
